+++ src/transform/toggle.ts
@@ -20,26 +20,23 @@ const implementWinningToggle = (
   const callPath = j(callExpression)
   const winningArgument = node.arguments[winnerArgumentIndex]
 
-  const notOfType = (type, node) => j(node).closest(type).size() === 0
-
-  const findUnusedReferencesOfType = (type, name) =>
-    root
-      .find(j.Identifier, {name})
-      .filter(notOfType.bind(null, type))
-      .size() === 0
+  // Looks for references only referred in the losing arrow function usage,
+  // and one declaration - hence 2 in total
+  const findUnusedReferencesOfType = (name) =>
+    root.find(j.Identifier, {name}).size() === 2
 
   const removeUnusedReferences = (type, name, identifiers) =>
     root
       .find(type, identifiers)
-      .filter(findUnusedReferencesOfType.bind(null, type, name))
+      .filter(findUnusedReferencesOfType.bind(null, name))
       .remove()
 
-  // Clean up dangling losing variable references
   const losingArgumentFunctions = node.arguments.filter(
     (arg, index) =>
       arg.type === 'ArrowFunctionExpression' && index !== winnerArgumentIndex
   )
 
+  // Clean up dangling losing variable references
   losingArgumentFunctions.forEach((losingFunction) => {
     j(losingFunction.body)
       .find(j.Identifier)
@@ -56,19 +53,25 @@ const implementWinningToggle = (
       })
   })
 
+  // if a parent is an JSX expression, remove it altogether
+  // <>{toggle('foo', <A/>, <B/>)}</> becomes <><B/></> and not <>{<B/>}<>
+  const parent = callExpression.parentPath
+  const parentIsJSXExpression = parent.value.type === 'JSXExpressionContainer'
+  const nodeToClean = parentIsJSXExpression ? j(parent) : callPath
+
   // Winner implementation
   // function, use body
   if (winningArgument.type === 'ArrowFunctionExpression') {
-    callPath.replaceWith(winningArgument.body)
+    nodeToClean.replaceWith(winningArgument.body)
   } else if (
     // null value, remove
-    winningArgument.type === 'Literal' &&
+    winningArgument.type === 'NullLiteral' &&
     winningArgument.value === null
   ) {
-    callPath.remove()
+    nodeToClean.remove()
   } else {
     // use raw value
-    callPath.replaceWith(node.arguments[winnerArgumentIndex])
+    nodeToClean.replaceWith(node.arguments[winnerArgumentIndex])
   }
 }